[refactor] merge autofunctions/focus.py into autofocus.py
